Update paid-search-detection.md
Updating the documentation based on recent client interactions with the documentation. The current version states that the rules "Specifies a string for a case-sensitive match with any part of the query string parameter". The parameters are the portions of a query string before the equal sign. For example, in this URL (https://test.dekalb.com.co/es-co.html?mktoprogramid=2345&mktocampaignid=9874&mktoleadid=4116733&utm_medium=testing), the query string parameters are 'mktoprogramid', 'mktocampaignid', 'mktoleadid', and 'utm_medium'. The query string is every character after the "?". 

Also, recommended adding a note that customers use query string parameters that are sufficiently distinct, since our current list of available operations do not easily support distinguishing similar string patterns.